refactor: split import pipeline step into modules
